requirements.txt
# MCQ Quiz Application - Requirements
# Python 3.8+

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
mysql-connector-python==8.2.0

# Security
bcrypt==4.1.2
Flask-WTF==1.2.1
python-dotenv==1.0.0

# Development (optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
